requirements.txt
ultralytics==8.0.202
Pillow==9.3.0
numpy==1.24.3
tensorflow==2.14.0
opencv-python==4.6.0.66
gdown==4.7.1